feat: 디테일 페이지까지 구현
