+++ ptn_project/views.py
@@ -3,8 +3,3 @@ from rest_framework import viewsets, mixins
 from .models import *
 from .serializers import *
 # Create your views here.
-class ProjectViewSet(viewsets.ModelViewSet):
-    queryset = Project.objects.all()
-    serializer_class = ProjectSerializer
-
-
